+++ Rdmp.Core/CohortCreation/Execution/CohortCompiler.cs
@@ -331,55 +331,63 @@ namespace Rdmp.Core.CohortCreation.Execution
 
         public void CacheSingleTask(ICacheableTask cacheableTask, ExternalDatabaseServer queryCachingServer)
         {
-            //if it is already cached don't inception cache
-            var sql = Tasks[cacheableTask].CountSQL;
-
-            if (sql.Trim().StartsWith(CachedAggregateConfigurationResultsManager.CachingPrefix))
-                return;
-            
-            var manager = new CachedAggregateConfigurationResultsManager(queryCachingServer);
-
-            var explicitTypes = new List<DatabaseColumnRequest>();
-
-            AggregateConfiguration configuration = cacheableTask.GetAggregateConfiguration();
             try
             {
-                //the identifier column that we read from
-                var identifiers = configuration.AggregateDimensions.Where(c => c.IsExtractionIdentifier).ToArray();
+                //if it is already cached don't inception cache
+                var sql = Tasks[cacheableTask].CountSQL;
 
-                if (identifiers.Length != 1)
-                    throw new Exception(string.Format(
-                        "There were {0} columns in the configuration marked IsExtractionIdentifier:{1}",
-                        identifiers.Length, string.Join(",", identifiers.Select(i => i.GetRuntimeName()))));
-
-                var identifierDimension = identifiers[0];
-                ColumnInfo identifierColumnInfo = identifierDimension.ColumnInfo;
-                var destinationDataType = GetDestinationType(identifierColumnInfo.Data_type,cacheableTask,queryCachingServer);
+                if (sql.Trim().StartsWith(CachedAggregateConfigurationResultsManager.CachingPrefix))
+                    return;
                 
-                explicitTypes.Add(new DatabaseColumnRequest(identifierDimension.GetRuntimeName(), destinationDataType));
+                var manager = new CachedAggregateConfigurationResultsManager(queryCachingServer);
 
-                //make other non transform Types have explicit values
-                foreach(AggregateDimension d in configuration.AggregateDimensions)
+                var explicitTypes = new List<DatabaseColumnRequest>();
+
+                AggregateConfiguration configuration = cacheableTask.GetAggregateConfiguration();
+                try
                 {
-                    if(d != identifierDimension)
+                    //the identifier column that we read from
+                    var identifiers = configuration.AggregateDimensions.Where(c => c.IsExtractionIdentifier).ToArray();
+
+                    if (identifiers.Length != 1)
+                        throw new Exception(string.Format(
+                            "There were {0} columns in the configuration marked IsExtractionIdentifier:{1}",
+                            identifiers.Length, string.Join(",", identifiers.Select(i => i.GetRuntimeName()))));
+
+                    var identifierDimension = identifiers[0];
+                    ColumnInfo identifierColumnInfo = identifierDimension.ColumnInfo;
+                    var destinationDataType = GetDestinationType(identifierColumnInfo.Data_type,cacheableTask,queryCachingServer);
+                    
+                    explicitTypes.Add(new DatabaseColumnRequest(identifierDimension.GetRuntimeName(), destinationDataType));
+
+                    //make other non transform Types have explicit values
+                    foreach(AggregateDimension d in configuration.AggregateDimensions)
                     {
-                        //if the user has not changed the SelectSQL and the SelectSQL of the original column is not a transform
-                        if(d.ExtractionInformation.SelectSQL.Equals(d.SelectSQL) && !d.ExtractionInformation.IsProperTransform())
+                        if(d != identifierDimension)
                         {
-                            //then use the origin datatype
-                            explicitTypes.Add(new DatabaseColumnRequest(d.GetRuntimeName(),GetDestinationType(d.ExtractionInformation.ColumnInfo.Data_type, cacheableTask, queryCachingServer)));
+                            //if the user has not changed the SelectSQL and the SelectSQL of the original column is not a transform
+                            if(d.ExtractionInformation.SelectSQL.Equals(d.SelectSQL) && !d.ExtractionInformation.IsProperTransform())
+                            {
+                                //then use the origin datatype
+                                explicitTypes.Add(new DatabaseColumnRequest(d.GetRuntimeName(),GetDestinationType(d.ExtractionInformation.ColumnInfo.Data_type, cacheableTask, queryCachingServer)));
+                            }
                         }
                     }
                 }
+                catch (Exception e)
+                {
+                    throw new Exception("Error occurred trying to find the data type of the identifier column when attempting to submit the result data table to the cache", e);
+                }
+
+                CacheCommitArguments args = cacheableTask.GetCacheArguments(sql, Tasks[cacheableTask].Identifiers, explicitTypes.ToArray());
+
+                manager.CommitResults(args);
             }
             catch (Exception e)
             {
-                throw new Exception("Error occurred trying to find the data type of the identifier column when attempting to submit the result data table to the cache", e);
+                cacheableTask.State = CompilationState.Crashed;
+                cacheableTask.CrashMessage = new Exception("Failed to cache results",e);
             }
-
-            CacheCommitArguments args = cacheableTask.GetCacheArguments(sql, Tasks[cacheableTask].Identifiers, explicitTypes.ToArray());
-
-            manager.CommitResults(args);
         }
 
         /// <summary>