+++ Rdmp.Core/QueryBuilding/QueryBuilder.cs
@@ -7,6 +7,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Text;
 using System.Text.RegularExpressions;
 using FAnsi.Discovery.QuerySyntax;
 using MapsDirectlyToDatabaseTable.Injection;
@@ -348,6 +349,11 @@ namespace Rdmp.Core.QueryBuilding
                 toReturn += "SET " + sqlParameter.ParameterName + "=" + sqlParameter.Value + ";" + Environment.NewLine;//its a regular value
             
             return toReturn;
+        }
+
+        public static string GetParameterDeclarationSQL(IEnumerable<ISqlParameter> sqlParameters)
+        {
+            return string.Join("", sqlParameters.Select(GetParameterDeclarationSQL));
         }
     }
 }