+++ Rdmp.UI/SimpleDialogs/ExceptionViewer.cs
@@ -21,7 +21,7 @@ namespace Rdmp.UI.SimpleDialogs
     {
         private readonly Exception _exception;
         
-        private ExceptionViewer(string title, string message, Exception exception):base(new WideMessageBoxArgs(title,message,exception.StackTrace??Environment.StackTrace,null,WideMessageBoxTheme.Exception))
+        private ExceptionViewer(string title, string message, Exception exception):base(new WideMessageBoxArgs(title,message,GetStackTrace(exception,Environment.StackTrace),null,WideMessageBoxTheme.Exception))
         {
             _exception = exception;
 
@@ -35,7 +35,27 @@ namespace Rdmp.UI.SimpleDialogs
                     _exception = aggregateException.InnerExceptions[0];
             }
         }
-        
+
+        /// <summary>
+        /// Returns the first stack trace from the <paramref name="exception"/> (including examining inner exceptions where stack trace is missing).
+        /// Otherwise returns <paramref name="ifNotFound"/>. 
+        /// </summary>
+        /// <param name="exception"></param>
+        /// <param name="ifNotFound"></param>
+        /// <returns></returns>
+        private static string GetStackTrace(Exception exception, string ifNotFound)
+        {
+            while (exception != null)
+            {
+                if (exception.StackTrace != null)
+                    return exception.StackTrace;
+
+                exception = exception.InnerException;
+            }
+
+            return ifNotFound;
+        }
+
         public static void Show(Exception exception, bool isModalDialog = true)
         {
             var longMessage = "";